Reorganize, lint imports, for indices.
