+++ src/pages/Entrepreneurship.jsx
@@ -17,8 +17,12 @@ import Hero from './Hero'
 import Key from '../components/modules/Key'
 import Footer from './Footer'
 
+// Index
+// import EntrepreneurshipIndex from '../components/indices/EntrepreneurshipIndex'
+
 // Data 
-import { Projects } from '../data/EntrepreneurshipProjects'
+// import { Projects } from '../data/EntrepreneurshipProjects'
+
 
 class Entrepreneurship extends Component {
   constructor(props) {
